tsrx: split the transform into focused modules
The monolithic transform.cts (~1340 lines) is now an orchestrator + planning core
(903 lines) over five focused sibling modules, each compiled .cts -> .cjs by the
widened tsconfig include glob and required by the others. Behaviour-preserving:
fixtures byte-identical, 80 tests green, extension bundle still builds (esbuild
inlines the siblings from index.cjs).

- transform-types.cts    — the data model (interfaces + Ts), types-only
- transform-operators.cts — operatorTables(ts)
- transform-program.cts  — compiler options, createProgramFromText, runtime
                           resolution. The transform fn is INJECTED (not imported)
                           so the cross-.tsrx recursion needn't import the
                           orchestrator back — breaks the would-be cycle.
- transform-emit.cts     — assembleOutput + the segment-mapping utilities
- transform-checker.cts  — the predicates as createChecker(ts, checker, sf, state).
                           The "sensible refactor along the way": the shared mutable
                           state the predicates touch (observableBindings, aliasInfo,
                           maybeEmptyBindings, nonCallableBindings, stalls) is now an
                           explicit CheckerState object passed by reference. The
                           orchestrator destructures the result so call sites are
                           unchanged.

transform.cts re-exports the public API unchanged, so every require('./transform.cjs')
consumer (index.cjs, language-plugin.cjs, the harness) is unaffected. Cross-module
imports use the explicit .cjs extension (bare specifiers don't resolve under node
moduleResolution); the .cjs files are committed build artifacts as before.
