+++ tsrx/ts-plugin/transform.cjs
@@ -11,152 +11,15 @@
 'use strict';
 Object.defineProperty(exports, "__esModule", { value: true });
 const path = require('node:path');
-const fs = require('node:fs');
-// The runtime module the generated code imports `render` from. We emit a relative
-// specifier from each .tsrx file's directory so it resolves in both the headless
-// harness and the editor.
-//
-// The runtime is located by walking up from the .tsrx file's OWN directory, not
-// relative to this module — the plugin is bundled into the VS Code extension's
-// node_modules, far from the source tree, so anchoring on `__dirname` emitted an
-// unresolvable `../../…/node_modules/runtime` specifier and collapsed every lifted
-// binding to `any` in the installed extension. `__dirname/../runtime.ts` is only a
-// last-resort fallback (covers the in-repo plugin running from source).
-const FALLBACK_RUNTIME_PATH = path.join(__dirname, '..', 'runtime.ts');
-function findRuntimeFile(baseDir) {
-    let dir = baseDir;
-    for (let i = 0; i < 16; i++) {
-        const candidate = path.join(dir, 'runtime.ts');
-        if (fs.existsSync(candidate))
-            return candidate;
-        const parent = path.dirname(dir);
-        if (parent === dir)
-            break;
-        dir = parent;
-    }
-    return null;
-}
-function relativeRuntimeSpecifier(baseDir) {
-    const runtime = findRuntimeFile(baseDir) || FALLBACK_RUNTIME_PATH;
-    let rel = path.relative(baseDir, runtime).replace(/\.ts$/, '').split(path.sep).join('/');
-    if (!rel.startsWith('.'))
-        rel = `./${rel}`;
-    return rel;
-}
-function getCompilerOptions(ts) {
-    return {
-        target: ts.ScriptTarget.ES2020,
-        module: ts.ModuleKind.ESNext,
-        moduleResolution: ts.ModuleResolutionKind.Bundler,
-        strict: true,
-        skipLibCheck: true,
-        noEmit: true,
-    };
-}
-// A Program that serves `text` for `fileName` and reads everything else (libs,
-// rxjs) from disk. Lets us type-check an unsaved editor buffer.
-//
-// Cross-.tsrx imports: a `.tsrx` file importing another `.tsrx` must see the
-// imported reactive values with REAL types — lifting is checker-driven, so an
-// `any` import would silently stop derivations lifting. So we resolve `./foo` to a
-// sibling `foo.tsrx` (when normal resolution finds nothing) and serve it as TS,
-// transformed on the fly. (No cycle handling yet; .tsrx import graphs are shallow.)
-function createProgramFromText(ts, fileName, text, options) {
-    const host = ts.createCompilerHost(options, true);
-    const isTsrx = (f) => typeof f === 'string' && f.endsWith('.tsrx');
-    const tsrxCache = new Map();
-    const tsrxCode = (f) => {
-        const src = fs.readFileSync(f, 'utf8');
-        const hit = tsrxCache.get(f);
-        if (hit && hit.src === src)
-            return hit.code;
-        const { code } = transformWithMappings(ts, src, path.dirname(f));
-        tsrxCache.set(f, { src, code });
-        return code;
-    };
-    const getSourceFile = host.getSourceFile.bind(host);
-    host.getSourceFile = (name, languageVersion, ...rest) => {
-        if (name === fileName)
-            return ts.createSourceFile(name, text, languageVersion, true);
-        if (isTsrx(name))
-            return ts.createSourceFile(name, tsrxCode(name), languageVersion, true, ts.ScriptKind.TS);
-        return getSourceFile(name, languageVersion, ...rest);
-    };
-    host.fileExists = f => (f === fileName ? true : ts.sys.fileExists(f));
-    const readFile = host.readFile.bind(host);
-    host.readFile = f => (f === fileName ? text : isTsrx(f) ? tsrxCode(f) : readFile(f));
-    const resolve = (name, containingFile) => {
-        const standard = ts.resolveModuleName(name, containingFile, options, host).resolvedModule;
-        if (standard)
-            return standard;
-        if (name.startsWith('.')) {
-            const candidate = path.resolve(path.dirname(containingFile), `${name}.tsrx`);
-            if (ts.sys.fileExists(candidate))
-                return { resolvedFileName: candidate, extension: ts.Extension.Ts };
-        }
-        return undefined;
-    };
-    if (ts.resolveModuleNameLiterals) {
-        host.resolveModuleNameLiterals = (literals, containingFile) => literals.map(lit => ({ resolvedModule: resolve(lit.text, containingFile) }));
-    }
-    else {
-        host.resolveModuleNames = (names, containingFile) => names.map((n) => resolve(n, containingFile));
-    }
-    return ts.createProgram([fileName], options, host);
-}
-// The operators we lift, keyed by SyntaxKind → the source token we re-emit.
-// Built from the injected `ts` (SyntaxKind values are only known at runtime).
-function operatorTables(ts) {
-    return {
-        // Binary operators we lift eagerly via combineLatest + map. All emit a
-        // primitive (number / boolean / string), so the result is never callable.
-        // NOTE: short-circuiting operators (&& || ??) are deliberately absent — they
-        // get lazy switchMap handling below, like the ternary.
-        LIFTABLE: {
-            [ts.SyntaxKind.PlusToken]: '+',
-            [ts.SyntaxKind.MinusToken]: '-',
-            [ts.SyntaxKind.AsteriskToken]: '*',
-            [ts.SyntaxKind.SlashToken]: '/',
-            [ts.SyntaxKind.PercentToken]: '%',
-            [ts.SyntaxKind.AsteriskAsteriskToken]: '**',
-            // Comparisons → boolean streams (these feed ternary/logical conditions).
-            [ts.SyntaxKind.LessThanToken]: '<',
-            [ts.SyntaxKind.GreaterThanToken]: '>',
-            [ts.SyntaxKind.LessThanEqualsToken]: '<=',
-            [ts.SyntaxKind.GreaterThanEqualsToken]: '>=',
-            [ts.SyntaxKind.EqualsEqualsEqualsToken]: '===',
-            [ts.SyntaxKind.ExclamationEqualsEqualsToken]: '!==',
-            [ts.SyntaxKind.EqualsEqualsToken]: '==',
-            [ts.SyntaxKind.ExclamationEqualsToken]: '!=',
-            // Bitwise → number streams.
-            [ts.SyntaxKind.AmpersandToken]: '&',
-            [ts.SyntaxKind.BarToken]: '|',
-            [ts.SyntaxKind.CaretToken]: '^',
-            [ts.SyntaxKind.LessThanLessThanToken]: '<<',
-            [ts.SyntaxKind.GreaterThanGreaterThanToken]: '>>',
-            [ts.SyntaxKind.GreaterThanGreaterThanGreaterThanToken]: '>>>',
-        },
-        // Prefix unary operators we lift via map. All emit a primitive.
-        LIFTABLE_UNARY: {
-            [ts.SyntaxKind.ExclamationToken]: '!',
-            [ts.SyntaxKind.MinusToken]: '-',
-            [ts.SyntaxKind.PlusToken]: '+',
-            [ts.SyntaxKind.TildeToken]: '~',
-        },
-        // Short-circuiting operators: lifted lazily via switchMap (like the ternary) so
-        // the right-hand stream is only subscribed when the left actually selects it.
-        LOGICAL: {
-            [ts.SyntaxKind.AmpersandAmpersandToken]: '&&',
-            [ts.SyntaxKind.BarBarToken]: '||',
-            [ts.SyntaxKind.QuestionQuestionToken]: '??',
-        },
-    };
-}
+const transform_operators_cjs_1 = require("./transform-operators.cjs");
+const transform_program_cjs_1 = require("./transform-program.cjs");
+const transform_emit_cjs_1 = require("./transform-emit.cjs");
+const transform_checker_cjs_1 = require("./transform-checker.cjs");
 function transformWithMappings(ts, sourceText, baseDir) {
-    const { LIFTABLE, LIFTABLE_UNARY, LOGICAL } = operatorTables(ts);
+    const { LIFTABLE, LIFTABLE_UNARY, LOGICAL } = (0, transform_operators_cjs_1.operatorTables)(ts);
     const fileName = path.join(baseDir, '__tsrx_virtual__.ts');
-    const options = getCompilerOptions(ts);
-    const program = createProgramFromText(ts, fileName, sourceText, options);
+    const options = (0, transform_program_cjs_1.getCompilerOptions)(ts);
+    const program = (0, transform_program_cjs_1.createProgramFromText)(ts, fileName, sourceText, options, transformWithMappings);
     const checker = program.getTypeChecker();
     const sf = program.getSourceFile(fileName);
     // ---------------------------------------------------------------------------
@@ -214,167 +77,10 @@ function transformWithMappings(ts, sourceText, baseDir) {
     // mapping over a lifted arg nests it: Observable<Observable<…>>). Surfaced as a
     // warning, since it type-checks and would otherwise be a silent footgun.
     const higherOrder = [];
-    // ---------------------------------------------------------------------------
-    // Checker predicates. The transform's questions about the source, answered by
-    // the real type checker plus the tracked-binding state above. Three layers:
-    //   - pure type queries (isObservableType, observableValueType, typeIsCallable, …)
-    //     read only a `ts.Type`;
-    //   - observability of an EXPRESSION (isObservableExpr, isObservableChain) also
-    //     consults the tracked bindings/aliases the checker can't see post-transform;
-    //   - structural recognizers (returnsComponent, isComponentMapCall) classify a
-    //     node by what it would produce.
-    // ---------------------------------------------------------------------------
-    const isObservableType = (type) => {
-        if (!type)
-            return false;
-        if (type.isUnion && type.isUnion())
-            return type.types.some(isObservableType);
-        return Boolean(type.getProperty && type.getProperty('subscribe') && type.getProperty('pipe'));
-    };
-    const isObservableExpr = (node) => (ts.isIdentifier(node) && (observableBindings.has(node.text) || aliasInfo.has(node.text))) ||
-        isObservableType(checker.getTypeAtLocation(node));
-    // Does parameter `argIndex` of this call accept an Observable? Two callers ask
-    // this for different reasons: an operator-style call whose stream argument must
-    // pass through verbatim (not be mapped per emission), and an EventHandler slot
-    // (`handler | Observable<handler>`) that takes a lifted handler stream (D2).
-    const paramAcceptsObservable = (callNode, argIndex) => {
-        const sig = checker.getResolvedSignature(callNode);
-        if (!sig)
-            return false;
-        const params = sig.getParameters();
-        if (!params.length)
-            return false;
-        const pSym = params[Math.min(argIndex, params.length - 1)];
-        return Boolean(pSym) && isObservableType(checker.getTypeOfSymbolAtLocation(pSym, callNode));
-    };
-    // The value type T of an Observable<T>/RenderObservable<T>, or undefined if the
-    // type isn't an observable. Reads the type's first type argument.
-    const observableValueType = (type) => {
-        if (!type)
-            return undefined;
-        if (type.isUnion && type.isUnion()) {
-            for (const t of type.types) {
-                const v = observableValueType(t);
-                if (v)
-                    return v;
-            }
-            return undefined;
-        }
-        if (!isObservableType(type))
-            return undefined;
-        let args;
-        try {
-            args = checker.getTypeArguments(type);
-        }
-        catch {
-            args = type.typeArguments;
-        }
-        return args && args.length ? args[0] : undefined;
-    };
-    // Does this expression evaluate to an observable that completes WITHOUT emitting
-    // (RxJS `EMPTY`, i.e. Observable<never>)? Used to spot the `cond ? x : EMPTY`
-    // filter idiom so we can flag the resulting binding as maybe-empty.
-    const emitsNever = (node) => {
-        const value = observableValueType(checker.getTypeAtLocation(node));
-        return Boolean(value && value.flags & ts.TypeFlags.Never);
-    };
-    // Record a "stall" warning when an operand fed into a combineLatest is a
-    // maybe-empty binding: combineLatest waits for every source's first emission, so a
-    // source that can stay EMPTY freezes the whole derived value until (if ever) it
-    // emits. The `cond ? x : EMPTY` filter idiom is fine standalone (as a child) but
-    // hazardous when combined — this teaches that exactly where it happens.
-    const noteStall = (node) => {
-        if (ts.isIdentifier(node) && maybeEmptyBindings.has(node.text)) {
-            stalls.push({ start: node.getStart(sf), length: node.getEnd() - node.getStart(sf), name: node.text });
-        }
-    };
-    // Would calling this expression be a boundary violation rather than a real
-    // "observable emitting a function"? True when the callee is a stream whose
-    // emitted value can't be called: a binding we lifted to a non-callable value,
-    // or a checker-visible observable whose T has no call signatures.
-    const calleeEmitsNonCallable = (node) => {
-        if (ts.isIdentifier(node) && nonCallableBindings.has(node.text))
-            return true;
-        const valueType = observableValueType(checker.getTypeAtLocation(node));
-        if (valueType)
-            return checker.getSignaturesOfType(valueType, ts.SignatureKind.Call).length === 0;
-        return false;
-    };
-    const typeIsCallable = (type) => type ? checker.getSignaturesOfType(type, ts.SignatureKind.Call).length > 0 : false;
-    // The return type of a callable type's first call signature, or undefined.
-    const returnTypeOf = (type) => {
-        const sigs = type ? checker.getSignaturesOfType(type, ts.SignatureKind.Call) : [];
-        return sigs.length ? sigs[0].getReturnType() : undefined;
-    };
-    // The RxJS Observable/Subject surface — members that operate on the STREAM
-    // itself, not the value it emits. Used only as a fallback for tracked bindings
-    // whose pre-transform type the checker can't see; checker-visible observables
-    // are asked directly. (A tracked binding emitting an object with a field named
-    // like one of these — e.g. `.value` — would be read as a stream op; rare.)
-    const STREAM_MEMBERS = new Set([
-        'subscribe', 'pipe', 'forEach', 'lift', 'toPromise', 'asObservable', 'source', 'operator',
-        'next', 'error', 'complete', 'value', 'getValue', 'closed', 'observed',
-        'hasError', 'thrownError', 'unsubscribe', 'observers',
-    ]);
-    // For `obj.member` where obj is a stream: does `member` target the emitted
-    // value (→ lift via map) rather than the stream API (→ leave, it's .pipe/.value
-    // etc.)? Ask the checker when obj is checker-visible as an observable; else
-    // fall back to the known RxJS surface.
-    const memberTargetsValue = (objNode, name) => {
-        const t = checker.getTypeAtLocation(objNode);
-        if (isObservableType(t))
-            return !t.getProperty(name);
-        return !STREAM_MEMBERS.has(name);
-    };
-    // Structural test for a DOM element type — RxFM components emit these.
-    const isElementValueType = (type) => Boolean(type && type.getProperty && type.getProperty('nodeType') &&
-        (type.getProperty('tagName') || type.getProperty('nodeName')));
-    // Does this function expression return a component (an Observable of a DOM
-    // element)? Excludes `any` (which a value-mapping callback gets, since its item
-    // param is untyped once `.map` fails to resolve on Observable). Reliable because
-    // component creators (Div, …) are typed regardless of the item param.
-    const returnsComponent = (fnNode) => {
-        const sig = checker.getSignatureFromDeclaration(fnNode);
-        if (!sig)
-            return false;
-        const ret = checker.getReturnTypeOfSignature(sig);
-        if (ret.flags & ts.TypeFlags.Any)
-            return false;
-        return isElementValueType(observableValueType(ret));
-    };
-    // Side-effect-free observability for a member/method/index chain: true when the
-    // chain roots in an observable (so the whole thing lifts to stay observable).
-    // Lets `items.filter(...).map(...)` be recognised, where the checker can't type
-    // the derived receiver directly.
-    const isObservableChain = (node) => {
-        if (!node)
-            return false;
-        if (isObservableExpr(node))
-            return true;
-        if (ts.isCallExpression(node) && ts.isPropertyAccessExpression(node.expression)) {
-            return isObservableChain(node.expression.expression);
-        }
-        if (ts.isPropertyAccessExpression(node) || ts.isElementAccessExpression(node)) {
-            return isObservableChain(node.expression);
-        }
-        if (ts.isParenthesizedExpression(node))
-            return isObservableChain(node.expression);
-        return false;
-    };
-    // Is this `obj.map(cb, key?)` where obj is a stream and cb returns a component?
-    // Such a call is the imperative way to render a list — lift it to
-    // mapToComponents (keyed reconciliation) rather than a naive array map.
-    const isComponentMapCall = (node) => {
-        if (!ts.isCallExpression(node) || node.arguments.length < 1)
-            return false;
-        const ex = node.expression;
-        if (!ts.isPropertyAccessExpression(ex) || ex.name.text !== 'map')
-            return false;
-        const cb = node.arguments[0];
-        if (!ts.isArrowFunction(cb) && !ts.isFunctionExpression(cb))
-            return false;
-        return isObservableChain(ex.expression) && returnsComponent(cb);
-    };
+    // The checker predicates (see transform-checker), bound to this run's checker and
+    // the slice of state above they read/write. Destructured so the planning code
+    // below reads them as plain predicates.
+    const { isObservableType, isObservableExpr, paramAcceptsObservable, observableValueType, emitsNever, noteStall, calleeEmitsNonCallable, typeIsCallable, returnTypeOf, memberTargetsValue, isElementValueType, isComponentMapCall, } = (0, transform_checker_cjs_1.createChecker)(ts, checker, sf, { observableBindings, aliasInfo, maybeEmptyBindings, nonCallableBindings, stalls });
     // The set of root identifier names of every observable VALUE referenced in `node`,
     // or null if the expression can't lift as one `x.pipe(map(x => <verbatim>))`. When
     // this is a single name, the whole expression lifts that way — it runs as plain code
@@ -1225,154 +931,10 @@ function transformWithMappings(ts, sourceText, baseDir) {
     // Assemble the planned edits into generated code + the source↔generated
     // position map. Everything above is the PLANNING phase (deciding what to
     // rewrite); this is the EMIT phase (splicing it together).
-    const { code, segments } = assembleOutput({
+    const { code, segments } = (0, transform_emit_cjs_1.assembleOutput)({
         ts, sf, sourceText, baseDir, edits, needed, usedRender, declMappings, rootMappings,
     });
     const sourceDiagnostics = ts.getPreEmitDiagnostics(program).filter(d => d.file && d.file.fileName === fileName);
     return { code, segments, sourceDiagnostics, stalls, higherOrder };
 }
-// Splice the planned edits over the source to produce the generated code, and
-// record the source↔generated SEGMENTS as we go. Pure with respect to the
-// transform's checker state — it consumes only the plan (edits + the import/
-// mapping bookkeeping) and the original text.
-function assembleOutput({ ts, sf, sourceText, baseDir, edits, needed, usedRender, declMappings, rootMappings }) {
-    // First generated position recorded per alias/root key, so the destructured-
-    // field declaration sites (D4) and single-root occurrences (D5) can map onto
-    // the `item.prop` / outer-stream token their first use emitted.
-    const aliasGenTarget = new Map();
-    // Names already imported per module — so we don't import a name twice. We do
-    // NOT modify the existing import statements: leaving them intact keeps full TS
-    // features (hover, go-to-def) on them via their identity mappings. The extra
-    // names we need go in a separate generated-only import line at the top (two
-    // imports from the same module is legal as long as names don't collide).
-    const alreadyImported = (mod) => {
-        const names = new Set();
-        for (const s of sf.statements) {
-            if (ts.isImportDeclaration(s) && ts.isStringLiteral(s.moduleSpecifier) && s.moduleSpecifier.text === mod) {
-                const nb = s.importClause && s.importClause.namedBindings;
-                if (nb && ts.isNamedImports(nb))
-                    for (const el of nb.elements)
-                        names.add(el.name.text);
-            }
-        }
-        return names;
-    };
-    const newNames = (mod, set) => [...set].filter(n => !alreadyImported(mod).has(n));
-    const importLines = [];
-    if (usedRender)
-        importLines.push(`import { render } from "${relativeRuntimeSpecifier(baseDir)}";`);
-    const rxjsNew = newNames('rxjs', needed.rxjs);
-    if (rxjsNew.length)
-        importLines.push(`import { ${rxjsNew.join(', ')} } from "rxjs";`);
-    const opNew = newNames('rxjs/operators', needed['rxjs/operators']);
-    if (opNew.length)
-        importLines.push(`import { ${opNew.join(', ')} } from "rxjs/operators";`);
-    const rxfmNew = newNames('rxfm', needed.rxfm);
-    if (rxfmNew.length)
-        importLines.push(`import { ${rxfmNew.join(', ')} } from "rxfm";`);
-    const importBlock = importLines.length ? importLines.join('\n') + '\n' : '';
-    edits.sort((a, b) => a.start - b.start);
-    let code = importBlock;
-    const segments = [];
-    let cursor = 0;
-    for (const edit of edits) {
-        if (edit.start > cursor) {
-            segments.push({ identity: true, srcStart: cursor, length: edit.start - cursor, genStart: code.length });
-            code += sourceText.slice(cursor, edit.start);
-        }
-        if (edit.remap !== undefined) {
-            // A source token mapped onto different generated text (e.g. `map` →
-            // `mapToComponents`). Navigable coarse mapping so hover/go-to-def land on
-            // the generated symbol; safe from smear since it's one token ↔ one token.
-            segments.push({ identity: false, navigable: true, srcStart: edit.start, srcLen: edit.end - edit.start, genStart: code.length, genLen: edit.remap.length });
-            code += edit.remap;
-            cursor = edit.end;
-            continue;
-        }
-        // Emit the rewritten initializer piece-by-piece: verbatim source slices get
-        // 1:1 identity segments (full language features); synthetic scaffolding is
-        // appended as generated-only text with no mapping.
-        for (const piece of edit.pieces) {
-            if (typeof piece === 'string') {
-                code += piece;
-            }
-            else if ('gen' in piece) {
-                // Generated-only text whose position is recorded as a mapping target (the
-                // outer stream reference of a single-root collapse — D5 root hover).
-                if (piece.refKey)
-                    aliasGenTarget.set(piece.refKey, { genStart: code.length, len: piece.gen.length });
-                code += piece.gen;
-            }
-            else {
-                const length = piece.srcEnd - piece.srcStart;
-                segments.push({ identity: true, srcStart: piece.srcStart, length, genStart: code.length });
-                // First generated `item.prop` token for this field: the declaration borrows it.
-                if (piece.aliasKey && !aliasGenTarget.has(piece.aliasKey))
-                    aliasGenTarget.set(piece.aliasKey, { genStart: code.length, len: length });
-                code += sourceText.slice(piece.srcStart, piece.srcEnd);
-            }
-        }
-        cursor = edit.end;
-    }
-    if (cursor < sourceText.length) {
-        segments.push({ identity: true, srcStart: cursor, length: sourceText.length - cursor, genStart: code.length });
-        code += sourceText.slice(cursor);
-    }
-    // Declaration-site mappings for destructured fields (D4): point each field's
-    // binding token at the generated `item.prop` token from its first use, so hover /
-    // go-to-def resolve on the binding too. These source spans sit in the erased
-    // binding-pattern region (no other segment covers them), so they never overlap.
-    for (const dm of declMappings) {
-        const target = aliasGenTarget.get(dm.aliasKey);
-        if (target)
-            segments.push({ identity: false, navigable: true, srcStart: dm.srcStart, srcLen: dm.len, genStart: target.genStart, genLen: target.len });
-    }
-    // D5 root hover: each single-root occurrence maps onto its collapse's outer stream
-    // reference, so the variable shows its `Observable<…>` type, not the in-map value.
-    for (const rm of rootMappings) {
-        const target = aliasGenTarget.get(rm.key);
-        if (target)
-            segments.push({ identity: false, navigable: true, srcStart: rm.srcStart, srcLen: rm.len, genStart: target.genStart, genLen: target.len });
-    }
-    return { code, segments };
-}
-function mapSourceToGenerated(segments, srcOffset) {
-    for (const seg of segments) {
-        if (seg.identity && srcOffset >= seg.srcStart && srcOffset < seg.srcStart + seg.length) {
-            return seg.genStart + (srcOffset - seg.srcStart);
-        }
-        if (!seg.identity && seg.srcLen != null && srcOffset >= seg.srcStart && srcOffset < seg.srcStart + seg.srcLen) {
-            return seg.genStart;
-        }
-    }
-    return null;
-}
-// Convert our segment table into Volar's CodeMapping[] format.
-//
-// Identity (1:1) segments — including the verbatim source tokens that survive
-// inside a rewritten expression — carry full language features. Synthesized
-// scaffolding is generated-only (no segment), so it never maps onto source
-// characters: nothing to smear. Non-identity segments map a source span onto
-// different generated text: `navigable` ones (a single token → single token,
-// e.g. `map` → `mapToComponents`) keep navigation + semantic so hover/go-to-def
-// work; the plain `coarse` profile (position-sensitive features off) is a
-// defensive fallback for any whole-span remap.
-function segmentsToVolarMappings(segments) {
-    const full = { completion: true, format: false, navigation: true, semantic: true, structure: true, verification: true };
-    const coarse = { completion: false, format: false, navigation: false, semantic: false, structure: false, verification: true };
-    const navigable = { completion: false, format: false, navigation: true, semantic: true, structure: false, verification: true };
-    const mappings = [];
-    for (const seg of segments) {
-        if (seg.identity) {
-            mappings.push({ sourceOffsets: [seg.srcStart], generatedOffsets: [seg.genStart], lengths: [seg.length], data: full });
-        }
-        else {
-            mappings.push({
-                sourceOffsets: [seg.srcStart], generatedOffsets: [seg.genStart],
-                lengths: [seg.srcLen], generatedLengths: [seg.genLen], data: seg.navigable ? navigable : coarse,
-            });
-        }
-    }
-    return mappings;
-}
-module.exports = { transformWithMappings, mapSourceToGenerated, segmentsToVolarMappings, getCompilerOptions };
+module.exports = { transformWithMappings, mapSourceToGenerated: transform_emit_cjs_1.mapSourceToGenerated, segmentsToVolarMappings: transform_emit_cjs_1.segmentsToVolarMappings, getCompilerOptions: transform_program_cjs_1.getCompilerOptions };
