+++ tsrx/ts-plugin/tsconfig.json
@@ -25,5 +25,5 @@
     "forceConsistentCasingInFileNames": true,
     "types": ["node"]
   },
-  "include": ["transform.cts"]
+  "include": ["*.cts"]
 }
